chore: mark migration RunPytohn elidable
When using a `RunPython` command in a migration, this command lives on
forever in the migrations code, but is never expected to be run again -
especially since there's no `reverse_code` param passed to restore
previous state, making them irreversible.

Marking them as `elibable=True` enables the `squashmigrations` command
to know that the operations contained in these statements should not
attempt to be carried forward in the event of a squash, rather the
engine will remove them completely.
This should only be done for model behavior that has been changed, or of
the operation applies strictly to live/prod data, and does not apply to
dev/testing.

Refs: https://docs.djangoproject.com/en/5.2/ref/migration-operations/#django.db.migrations.operations.RunPython
Refs: https://docs.djangoproject.com/en/5.2/topics/migrations/#migration-squashing

Signed-off-by: Mike Fiedler <miketheman@gmail.com>