test: add user dashboard tests
Demonstrate more behave testing of stave-specific features.
Multiple scenarios with many steps to demonstrate some of the
expressiveness available.

Left notes to refactor to Factory patterns when we have test factories.
While these tests are indeed slower, as they bootstrap an entire
application context and call with a web test client, the scenarios help
describe the overall intent of a feature flow.

Signed-off-by: Mike Fiedler <miketheman@gmail.com>